projects/ai-invoice-intelligence/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.5.0",
    "fastapi>=0.128.7",
    "openai>=2.20.0",
    "pillow>=12.1.0",
    "pydantic>=2.12.5",
    "pypdf>=6.7.0",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "sqlalchemy>=2.0.46",
    "uvicorn>=0.40.0",
]
